Refactor aroma-euro parser with improved parsing and brand detection logic
