+++ FINAL_VERSION_INFO.md
@@ -1,0 +1,67 @@
+# Финальная версия проекта
+
+## Анализ версий
+
+После анализа двух версий парсера было установлено:
+
+### Удаленная версия: `aroma_euro_final.py` (версия 3.0)
+- Более новая, но упрощенная версия
+- Менее функциональная логика поиска товаров
+- Простая обработка пагинации
+- Базовый функционал
+
+### Сохраненная версия: `aroma_euro_final.py` (версия final-2.0)
+- Более функциональная и надежная версия
+- Сложная логика поиска товаров с множественными селекторами
+- Детальная обработка пагинации
+- Расширенная обработка ошибок
+- Более точное разделение брендов и названий
+
+## Финальная структура проекта
+
+```
+├── aroma_euro_final.py          # Основной парсер (версия final-2.0)
+├── aroma_euro_perfumes_final.json  # Финальные данные (328 записей)
+├── README.md                    # Обновленная документация
+├── requirements.txt             # Зависимости проекта
+└── LICENSE                      # Лицензия проекта
+```
+
+## Особенности финальной версии
+
+### Расширенная логика поиска товаров:
+- Множественные CSS селекторы для поиска товаров
+- Альтернативные методы поиска по ссылкам
+- Группировка по родительским элементам
+- Обработка различных структур страниц
+
+### Улучшенная обработка пагинации:
+- Поиск по различным селекторам пагинации
+- Обработка числовых ссылок
+- Сортировка найденных страниц
+
+### Детальная обработка ошибок:
+- Сохранение отладочных HTML файлов
+- Подробное логирование
+- Множественные попытки поиска
+
+### Точное разделение брендов:
+- Расширенный список известных брендов
+- Сложные регулярные выражения
+- Обработка различных форматов названий
+
+## Запуск проекта
+
+```bash
+# Установка зависимостей
+pip install -r requirements.txt
+
+# Запуск парсера
+python aroma_euro_final.py
+```
+
+## Примечание
+
+Текущая версия содержит базовую информацию о товарах (название, бренд, цена, URL). 
+Расширенная информация о фабрике, артикуле, поле и т.д. не была найдена в доступных версиях.
+Возможно, эта функциональность была в одном из удаленных промежуточных файлов.
