+++ aroma_euro_final.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 """
-Финальная версия парсера каталога парфюмерных композиций с сайта aroma-euro.ru
+Специализированный парсер каталога парфюмерных композиций с сайта aroma-euro.ru
 Извлекает названия и бренды парфюмерии и сохраняет в JSON формате
 """
 
@@ -10,7 +10,7 @@ from bs4 import BeautifulSoup
 import json
 import time
 import logging
-from urllib.parse import urljoin, parse_qs, urlparse
+from urllib.parse import urljoin, urlparse, parse_qs
 import re
 from typing import List, Dict, Optional
 import sys
@@ -20,15 +20,15 @@ logging.basicConfig(
     level=logging.INFO,
     format='%(asctime)s - %(levelname)s - %(message)s',
     handlers=[
-        logging.FileHandler('aroma_euro_final.log', encoding='utf-8'),
+        logging.FileHandler('aroma_euro_parser.log', encoding='utf-8'),
         logging.StreamHandler()
     ]
 )
 logger = logging.getLogger(__name__)
 
 
-class AromaEuroFinalParser:
-    """Финальная версия парсера для сайта aroma-euro.ru"""
+class AromaEuroParser:
+    """Специализированный парсер для сайта aroma-euro.ru"""
     
     def __init__(self, base_url: str = "https://aroma-euro.ru"):
         self.base_url = base_url
@@ -60,108 +60,52 @@ class AromaEuroFinalParser:
             logger.error(f"Неожиданная ошибка при обработке {url}: {e}")
             return None
     
-    def parse_title(self, title: str) -> tuple:
-        """Парсинг названия для разделения на бренд и название"""
-        title = title.strip()
-        
-        # Удаляем лишние части
-        title = re.sub(r'\s*\(мотив\)\s*', '', title)
-        title = re.sub(r',\s*[A-Z]+\s*$', '', title)  # Убираем код в конце
-        
-        # Паттерны для разделения
-        patterns = [
-            # Tom Ford Black Orchid
-            r'^([A-Za-z][A-Za-z\s&\.\-\']+?)\s+([A-Za-z].+)$',
-            # Brand - Name
-            r'^([A-Za-z][A-Za-z\s&\.\-\']+?)\s*[-–—]\s*(.+)$',
-        ]
-        
-        for pattern in patterns:
-            match = re.match(pattern, title, re.IGNORECASE)
-            if match:
-                potential_brand = match.group(1).strip()
-                potential_name = match.group(2).strip()
-                
-                # Проверяем, что это разумное разделение
-                if (len(potential_brand.split()) <= 3 and 
-                    len(potential_brand) <= 50 and
-                    len(potential_name) > 0):
-                    return potential_brand, potential_name
-        
-        # Известные бренды
-        known_brands = [
-            'Tom Ford', 'Chanel', 'Dior', 'Gucci', 'Prada', 'Versace', 'Armani', 
-            'Calvin Klein', 'Hugo Boss', 'Dolce Gabbana', 'Yves Saint Laurent',
-            'Creed', 'Maison Margiela', 'Byredo', 'Le Labo', 'Diptyque', 'Hermès',
-            'Bottega Veneta', 'Burberry', 'Givenchy', 'Lancome', 'Estee Lauder',
-            'Escentric Molecules', 'Montale', 'Mancera', 'Amouage', 'Nasomatto',
-            'Serge Lutens', 'Francis Kurkdjian', 'Frederic Malle', 'Penhaligons',
-            'Miller Harris', 'Acqua di Parma', 'Annick Goutal', 'L\'Artisan Parfumeur'
-        ]
-        
-        title_lower = title.lower()
-        for brand in known_brands:
-            if title_lower.startswith(brand.lower()):
-                remaining = title[len(brand):].strip()
-                if remaining:
-                    return brand, remaining
-        
-        # Если не удалось разделить
-        return "", title
-    
     def extract_perfume_info(self, product_element) -> Optional[Dict[str, str]]:
         """Извлечение информации о парфюме из элемента товара"""
         try:
             perfume_info = {}
             
-            # Поиск названия товара
-            title_element = product_element.select_one('.product-title')
+            # Поиск названия товара - специфичные селекторы для aroma-euro.ru
+            title_element = None
             
+            # Попробуем различные селекторы для названия
+            title_selectors = [
+                '.ty-grid-list__item-name a',
+                '.product-title',
+                '.ty-product-title',
+                'a[class*="product-title"]',
+                '.ut2-gl__name a',
+                'h3 a',
+                'h4 a',
+                'a[href*="/perfume/"]'
+            ]
+            
+            for selector in title_selectors:
+                title_element = product_element.select_one(selector)
+                if title_element:
+                    break
+            
+            # Альтернативный поиск
             if not title_element:
-                # Альтернативные селекторы
-                title_selectors = [
-                    'a[href*="/perfume/"][href$="/"]',
-                    'a[href*="/perfume/"]',
-                    '.ty-grid-list__item-name a',
-                    '.ut2-gl__name a',
-                    'h3 a',
-                    'h4 a'
-                ]
-                
-                for selector in title_selectors:
-                    title_element = product_element.select_one(selector)
-                    if title_element:
-                        break
+                title_element = product_element.find('a', href=re.compile(r'.*/perfume/.*'))
             
             if title_element:
                 title_text = title_element.get_text(strip=True)
-                
-                # Фильтруем нерелевантные заголовки
-                if any(skip in title_text.lower() for skip in [
-                    'парфюмерные масла', 'весь каталог', 'моно ароматы', 
-                    'флаконы', 'аксессуары', 'хиты продаж', 'бренды', 'новинки'
-                ]):
-                    return None
-                
-                if title_text and len(title_text) > 5:
+                if title_text:
                     perfume_info['full_title'] = title_text
                     
-                    # Разделение на бренд и название
+                    # Попытка разделить название и бренд
                     brand, name = self.parse_title(title_text)
                     perfume_info['brand'] = brand
                     perfume_info['name'] = name
                     
-                    # URL товара
+                    # Поиск ссылки на товар
                     if title_element.get('href'):
-                        href = title_element['href']
-                        # Проверяем, что это ссылка на конкретный товар
-                        if '/perfume/' in href and not href.endswith('/perfume/'):
-                            perfume_info['url'] = urljoin(self.base_url, href)
-                        else:
-                            return None
+                        perfume_info['url'] = urljoin(self.base_url, title_element['href'])
                 else:
                     return None
             else:
+                logger.warning("Не найден элемент с названием товара")
                 return None
             
             # Поиск цены
@@ -183,15 +127,58 @@ class AromaEuroFinalParser:
             # Поиск изображения
             img_element = product_element.find('img')
             if img_element and img_element.get('src'):
-                src = img_element['src']
-                if 'product' in src.lower() or 'perfume' in src.lower():
-                    perfume_info['image'] = urljoin(self.base_url, src)
+                perfume_info['image'] = urljoin(self.base_url, img_element['src'])
             
             return perfume_info
             
         except Exception as e:
             logger.error(f"Ошибка при извлечении информации о товаре: {e}")
             return None
+    
+    def parse_title(self, title: str) -> tuple:
+        """Парсинг названия для разделения на бренд и название"""
+        title = title.strip()
+        
+        # Специфичные паттерны для парфюмерии
+        patterns = [
+            # Tom Ford - Black Orchid
+            r'^([A-Za-z][A-Za-z\s&\.\-\']+?)\s*[-–—]\s*(.+)$',
+            # Chanel No 5 (бренд в начале)
+            r'^(Chanel|Dior|Gucci|Prada|Versace|Armani|Calvin Klein|Hugo Boss|Dolce Gabbana|Yves Saint Laurent|Tom Ford|Creed|Maison Margiela|Byredo|Le Labo|Diptyque|Hermès|Bottega Veneta|Burberry|Givenchy|Lancome|Estee Lauder|Clinique|Marc Jacobs|Viktor Rolf|Jean Paul Gaultier|Thierry Mugler|Issey Miyake|Kenzo|Cacharel|Davidoff|Montblanc|Bvlgari|Cartier|Chopard|Tiffany|Van Cleef Arpels|Acqua di Parma|Annick Goutal|Penhaligons|Miller Harris|Escentric Molecules|Nasomatto|Amouage|Serge Lutens|Francis Kurkdjian|Frederic Malle|L\'Artisan Parfumeur|Diptyque|Comme des Garcons|Malin Goetz|Fresh|Philosophy|The Body Shop|Lush|Bath Body Works|Victoria\'s Secret|Abercrombie Fitch|Hollister|American Eagle|Aeropostale|Tommy Hilfiger|Ralph Lauren|Polo|Lacoste|Nautica|Perry Ellis|Kenneth Cole|Michael Kors|Coach|Kate Spade|Tory Burch|Donna Karan|DKNY|Vera Wang|Oscar de la Renta|Carolina Herrera|Narciso Rodriguez|Alexander McQueen|Stella McCartney|Vivienne Westwood|John Varvatos|Zegna|Acqua di Gio|Acqua di Gioia|1 Million|Invictus|Sauvage|Bleu de Chanel|Allure|Coco|Chance|Black Opium|Libre|Mon Paris|La Vie Est Belle|Tresor|Miracle|Hypnose|Obsession|Eternity|Escape|CK One|CK Be|Euphoria|Deep Euphoria|Guilty|Flora|Bamboo|Rush|Dylan Blue|Eros|Pour Homme|Femme|Code|Si|Because It\'s You|Stronger With You|The One|Light Blue|Dolce|Imperatrice|Good Girl|Bad Boy|212|Carolina Herrera|Alien|Angel|A Men|Pure Malt|Womanity|Aura|Mugler|L\'Eau d\'Issey|Pleats Please|Nuit d\'Issey|Flower|Jungle|World|Amor Amor|Yes I Am|Anais Anais|Lou Lou|Lolita Lempicka|Cool Water|Hot Water|Silver|Champion|Echo|Zino|White Tea|Silver Mountain Water|Aventus|Green Irish Tweed|Millisime Imperial|Royal Oud|Original Santal|Virgin Island Water|Love in White|Sublime Vanille|Tabarome|Erolfa|Bois du Portugal|Himalaya|Aberdeen Lavender|Royal Princess Oud|Fleurs de Bulgarie|Rose Goldea|Omnia|Serpenti|Man|Aqva|Tygar|Declaration|Terre d\'Hermes|Un Jardin|Voyage|Kelly Caleche|Twilly|H24|Equipage|Amazone|Calyx|Pleasures|White Linen|Beautiful|Youth Dew|Azuree|Private Collection|Sensuous|Bronze Goddess|Intuition|Knowing|Spellbound|Tuscany|White Tea|Daisy|Perfect|Decadence|Honey|Dot|Lola|Flowerbomb|Spicebomb|Good Fortune|Bonbon|Antidote|Only the Brave|Diesel|Fuel for Life|Bad|Loverdose|Sound of the Brave|Spirit|D&G|The One|Pour Femme|Intenso|K|Masculine|Garden|Anthology|Velvet|Rose The One|Royal Night|Devotion|Light Blue Forever|Imperatrice Limited Edition|Sicily|Carnal Flower|Portrait of a Lady|Iris Poudre|Lys Mediterranee|Dans Tes Bras|Une Fleur de Cassie|Bigarade Concentree|Carnal Flower|Dominique Ropion|Noir Epices|Outrageous|Russian Tea|Dries Van Noten|Vetiver Extraordinaire|Bois d\'Argent|Oud Wood|Tobacco Vanille|Neroli Portofino|Lost Cherry|Bitter Peach|Ebene Fume|Oud Minerale|Soleil Blanc|White Suede|Grey Vetiver|Mandarino di Amalfi|Costa Azzurra|Metallique|Ombre Leather|Fucking Fabulous|Tuscan Leather|Champaca Absolute|Plum Japonais|Cafe Rose|Shanghai Lily|Jasmin Rouge|Velvet Orchid|Black Orchid|White Patchouli|Vanille Fatale|Santal Blush|Electric Cherry|Rose Prick|Lavender Extreme|Vert Boheme|Vert d\'Encens|Beau de Jour|Vert de Fleur|Sole di Positano|Fleur de Portofino|Venetian Bergamot|Azure Lime|Acqua Essenziale|Colonia|Colonia Intensa|Colonia Oud|Colonia Pura|Iris Nobile|Peonia Nobile|Rosa Nobile|Magnolia Nobile|Vaniglia del Madagascar|Sandalo|Mirra|Ambra|Oud|Leather|Tobacco|Vetiver|Patchouli|Sandalwood|Cedar|Bergamot|Lemon|Orange|Grapefruit|Mandarin|Lime|Yuzu|Pink Pepper|Black Pepper|Cardamom|Coriander|Ginger|Nutmeg|Cinnamon|Clove|Star Anise|Juniper|Pine|Fir|Eucalyptus|Mint|Basil|Rosemary|Thyme|Lavender|Sage|Geranium|Rose|Jasmine|Tuberose|Ylang Ylang|Neroli|Orange Blossom|Lily|Lily of the Valley|Peony|Iris|Violet|Freesia|Gardenia|Magnolia|Honeysuckle|Mimosa|Osmanthus|Narcissus|Hyacinth|Cyclamen|Peach|Apple|Pear|Plum|Cherry|Strawberry|Raspberry|Blackberry|Blackcurrant|Coconut|Pineapple|Mango|Papaya|Passion Fruit|Fig|Date|Raisin|Prune|Vanilla|Caramel|Honey|Chocolate|Coffee|Cocoa|Almond|Hazelnut|Pistachio|Walnut|Chestnut|Praline|Toffee|Butterscotch|Marshmallow|Cotton Candy|Bubblegum|Candy|Sugar|Cream|Milk|Butter|Cheese|Bread|Cake|Cookie|Pie|Pastry|Doughnut|Waffle|Pancake|Syrup|Jam|Jelly|Marmalade|Compote|Sorbet|Ice Cream|Yogurt|Custard|Pudding|Mousse|Souffle|Meringue|Macaron|Madeleine|Croissant|Baguette|Brioche|Pain au Chocolat|Eclair|Profiterole|Chouquette|Paris Brest|Saint Honore|Opera|Mille Feuille|Tarte Tatin|Clafoutis|Creme Brulee|Flan|Tiramisu|Panna Cotta|Gelato|Granita|Affogato|Cappuccino|Espresso|Latte|Macchiato|Mocha|Hot Chocolate|Tea|Earl Grey|English Breakfast|Green Tea|Oolong|Pu Erh|White Tea|Rooibos|Chamomile|Peppermint|Spearmint|Ginseng|Matcha|Chai|Masala Chai|Turkish Tea|Russian Tea|Moroccan Tea|Indian Tea|Chinese Tea|Japanese Tea|Ceylon Tea|Assam|Darjeeling|Lapsang Souchong|Jasmine Tea|Rose Tea|Chrysanthemum Tea|Hibiscus Tea|Elderflower|Linden|Verbena|Lemon Balm|Fennel|Anise|Licorice|Cardamom Tea|Cinnamon Tea|Ginger Tea|Turmeric Tea|Kombucha|Kefir|Wine|Champagne|Prosecco|Cava|Sake|Beer|Whiskey|Bourbon|Scotch|Rum|Vodka|Gin|Tequila|Brandy|Cognac|Armagnac|Calvados|Grappa|Ouzo|Raki|Absinthe|Pastis|Sambuca|Limoncello|Amaretto|Baileys|Kahlua|Cointreau|Grand Marnier|Chartreuse|Benedictine|Drambuie|Frangelico|Galliano|Jagermeister|Sambuca|Strega|Aperol|Campari|Cynar|Fernet|Punt e Mes|Vermouth|Sherry|Port|Madeira|Marsala|Moscato|Riesling|Gewurztraminer|Pinot Grigio|Pinot Noir|Chardonnay|Sauvignon Blanc|Merlot|Cabernet Sauvignon|Syrah|Shiraz|Grenache|Sangiovese|Chianti|Barolo|Barbaresco|Brunello|Amarone|Prosecco|Franciacorta|Asti|Lambrusco|Soave|Valpolicella|Montepulciano|Nero d\'Avola|Primitivo|Aglianico|Fiano|Falanghina|Greco|Vermentino|Pecorino|Passerina|Trebbiano|Malvasia|Moscato d\'Asti|Brachetto|Dolcetto|Barbera|Nebbiolo|Cortese|Arneis|Roero|Gavi|Orvieto|Frascati|Est Est Est|Cesanese|Montepulciano d\'Abruzzo|Sangiovese di Romagna|Lambrusco di Sorbara|Lambrusco Grasparossa|Lambrusco Salamino|Pignoletto|Albana|Trebbiano d\'Abruzzo|Pecorino d\'Abruzzo|Montepulciano del Molise|Taurasi|Greco di Tufo|Fiano di Avellino|Lacryma Christi|Falanghina del Sannio|Aglianico del Taburno|Piedirosso|Sciascinoso|Coda di Volpe|Biancolella|Forastera|San Lunardo|Caprettone|Ginestra|Olivella|Ripolo|Uva Rara|Vespolina|Bonarda|Croatina|Ruchè|Grignolino|Freisa|Brachetto d\'Acqui|Moscato Rosa|Gewürztraminer|Sylvaner|Müller Thurgau|Pinot Bianco|Pinot Nero|Schiava|Lagrein|Teroldego|Marzemino|Nosiola|Chardonnay|Sauvignon|Merlot|Cabernet|Refosco|Friulano|Ribolla Gialla|Malvasia Istriana|Vitovska|Glera|Prosecco di Valdobbiadene|Prosecco di Conegliano|Cartizze|Rive|Millesimato|Brut|Extra Brut|Pas Dosé|Extra Dry|Dry|Demi Sec|Dolce|Spumante|Frizzante|Tranquillo|Novello|Riserva|Superiore|Classico|DOCG|DOC|IGT|Vino da Tavola|Biologico|Biodinamico|Naturale|Orange Wine|Pet Nat|Ancestrale|Col Fondo|Rifermentato|Charmat|Martinotti|Champenoise|Metodo Classico|Cuvée|Assemblage|Blend|Monovitigno|Cru|Vigna|Vigneto|Terroir|Millésime|Annata|Vendemmia|Harvest|Grape|Uva|Vine|Vite|Vineyard|Vignoble|Winery|Cantina|Cave|Cellar|Barrel|Botte|Tonneau|Barrique|Oak|Rovere|French Oak|American Oak|Slavonian Oak|Acacia|Cherry|Ciliegio|Chestnut|Castagno|Steel|Acciaio|Concrete|Cemento|Amphora|Anfora|Qvevri|Tinaja|Foudre|Puncheon|Hogshead|Firkin|Pin|Tun|Vat|Tank|Serbatoio|Fermentation|Fermentazione|Maceration|Macerazione|Malolactic|Malolattica|Lees|Fecce|Batonnage|Riddling|Remuage|Disgorgement|Sboccatura|Dosage|Liqueur|Expedition|Tirage|Prise de Mousse|Autolysis|Autolisi|Aging|Invecchiamento|Maturation|Maturazione|Evolution|Evoluzione|Peak|Picco|Decline|Declino|Drinking Window|Finestra di Bevibilità|Cellar|Cantina|Storage|Conservazione|Temperature|Temperatura|Humidity|Umidità|Light|Luce|Vibration|Vibrazione|Cork|Sughero|Screwcap|Tappo a Vite|Synthetic|Sintetico|Glass|Vetro|Bottle|Bottiglia|Magnum|Double Magnum|Jeroboam|Rehoboam|Methuselah|Salmanazar|Balthazar|Nebuchadnezzar|Melchior|Solomon|Sovereign|Primat|Melchizedek)\s+(.+)$',
+            # Brand Name (in parentheses)
+            r'^(.+?)\s*\(([^)]+)\).*$',
+            # Brand space Name
+            r'^([A-Z][a-z]+(?:\s+[A-Z][a-z]+)*)\s+(.+)$',
+        ]
+        
+        for pattern in patterns:
+            match = re.match(pattern, title, re.IGNORECASE)
+            if match:
+                brand = match.group(1).strip()
+                name = match.group(2).strip()
+                
+                # Проверяем, что бренд разумной длины
+                if len(brand.split()) <= 4 and len(brand) <= 50:
+                    return brand, name
+        
+        # Если не удалось разделить, пытаемся найти известные бренды в начале
+        known_brands = [
+            'Tom Ford', 'Chanel', 'Dior', 'Gucci', 'Prada', 'Versace', 'Armani', 
+            'Calvin Klein', 'Hugo Boss', 'Dolce Gabbana', 'Yves Saint Laurent',
+            'Creed', 'Maison Margiela', 'Byredo', 'Le Labo', 'Diptyque', 'Hermès',
+            'Bottega Veneta', 'Burberry', 'Givenchy', 'Lancome', 'Estee Lauder'
+        ]
+        
+        title_lower = title.lower()
+        for brand in known_brands:
+            if title_lower.startswith(brand.lower()):
+                remaining = title[len(brand):].strip()
+                if remaining.startswith('-') or remaining.startswith('–') or remaining.startswith('—'):
+                    remaining = remaining[1:].strip()
+                return brand, remaining if remaining else title
+        
+        # Если ничего не найдено, возвращаем пустой бренд и полное название
+        return "", title
     
     def get_all_pages_urls(self) -> List[str]:
         """Получение ссылок на все страницы каталога"""
@@ -201,24 +188,24 @@ class AromaEuroFinalParser:
         if not soup:
             return urls
         
-        # Поиск пагинации
-        pagination_links = soup.find_all('a', href=re.compile(r'.*page-\d+.*'))
+        # Поиск пагинации для aroma-euro.ru
+        pagination_selectors = [
+            '.ty-pagination__item a',
+            '.pagination a',
+            'a[href*="page-"]',
+            '.ty-pagination a'
+        ]
         
-        for link in pagination_links:
-            href = link.get('href')
-            if href:
-                full_url = urljoin(self.base_url, href)
-                if full_url not in urls:
-                    urls.append(full_url)
-        
-        # Также ищем числовые ссылки пагинации
-        page_numbers = soup.find_all('a', string=re.compile(r'^\d+$'))
-        for link in page_numbers:
-            href = link.get('href')
-            if href and 'page' in href:
-                full_url = urljoin(self.base_url, href)
-                if full_url not in urls:
-                    urls.append(full_url)
+        for selector in pagination_selectors:
+            page_links = soup.select(selector)
+            if page_links:
+                for link in page_links:
+                    href = link.get('href')
+                    if href and ('page-' in href or 'page=' in href):
+                        full_url = urljoin(self.base_url, href)
+                        if full_url not in urls:
+                            urls.append(full_url)
+                break
         
         logger.info(f"Найдено страниц для парсинга: {len(urls)}")
         return sorted(urls)
@@ -231,37 +218,60 @@ class AromaEuroFinalParser:
         
         perfumes = []
         
-        # Ищем все элементы с классом product-title
-        product_titles = soup.find_all(class_='product-title')
-        logger.info(f"Найдено элементов .product-title: {len(product_titles)}")
+        # Специфичные селекторы для aroma-euro.ru
+        product_selectors = [
+            '.ty-grid-list__item',
+            '.ut2-gl__item',
+            '.ty-column4',
+            '.product-item',
+            'div[class*="grid"]',
+            'div[class*="product"]',
+            'div[class*="item"]'
+        ]
         
-        # Для каждого заголовка ищем родительский контейнер
-        processed_urls = set()
+        products = []
+        for selector in product_selectors:
+            found_products = soup.select(selector)
+            if found_products:
+                products = found_products
+                logger.info(f"Найдено товаров с селектором '{selector}': {len(products)}")
+                break
         
-        for title_element in product_titles:
-            # Находим родительский контейнер товара
-            product_container = title_element.find_parent(['div', 'li', 'article'])
-            if product_container:
-                perfume_info = self.extract_perfume_info(product_container)
-                if perfume_info and perfume_info.get('url'):
-                    # Избегаем дубликатов по URL
-                    if perfume_info['url'] not in processed_urls:
-                        processed_urls.add(perfume_info['url'])
-                        perfumes.append(perfume_info)
+        # Если основные селекторы не сработали, попробуем найти товары по ссылкам
+        if not products:
+            # Ищем все ссылки на товары парфюмерии
+            product_links = soup.find_all('a', href=re.compile(r'.*/perfume/[^/]+/?$'))
+            if product_links:
+                # Группируем по родительским элементам
+                parents = set()
+                for link in product_links:
+                    parent = link.find_parent(['div', 'li', 'article'])
+                    if parent:
+                        parents.add(parent)
+                products = list(parents)
+                logger.info(f"Найдено товаров по ссылкам: {len(products)}")
         
-        logger.info(f"Извлечено уникальных парфюмов со страницы {url}: {len(perfumes)}")
+        if not products:
+            logger.warning(f"Не найдено товаров на странице: {url}")
+            # Попробуем сохранить HTML для анализа
+            with open('debug_page.html', 'w', encoding='utf-8') as f:
+                f.write(soup.prettify())
+            logger.info("HTML страницы сохранен в debug_page.html для анализа")
+        
+        for product in products:
+            perfume_info = self.extract_perfume_info(product)
+            if perfume_info:
+                perfumes.append(perfume_info)
+        
+        logger.info(f"Извлечено парфюмов со страницы {url}: {len(perfumes)}")
         return perfumes
     
-    def parse_all_catalog(self, max_pages: int = None) -> List[Dict[str, str]]:
+    def parse_all_catalog(self) -> List[Dict[str, str]]:
         """Парсинг всего каталога"""
         logger.info("Начинаю парсинг каталога парфюмерии aroma-euro.ru")
         
         # Получаем все страницы каталога
         page_urls = self.get_all_pages_urls()
-        
-        if max_pages:
-            page_urls = page_urls[:max_pages]
-            logger.info(f"Ограничиваю парсинг до {max_pages} страниц")
         
         all_perfumes = []
         
@@ -273,33 +283,39 @@ class AromaEuroFinalParser:
             
             # Задержка между запросами
             if i < len(page_urls):
-                time.sleep(2)
+                time.sleep(3)  # Увеличиваем задержку для надежности
         
-        # Удаление дубликатов по URL
+        # Удаление дубликатов
         unique_perfumes = []
+        seen_titles = set()
         seen_urls = set()
         
         for perfume in all_perfumes:
+            title = perfume.get('full_title', '')
             url = perfume.get('url', '')
+            
+            # Используем URL как основной критерий уникальности
             if url and url not in seen_urls:
                 seen_urls.add(url)
+                unique_perfumes.append(perfume)
+            elif not url and title and title not in seen_titles:
+                seen_titles.add(title)
                 unique_perfumes.append(perfume)
         
         logger.info(f"Всего найдено уникальных парфюмов: {len(unique_perfumes)}")
         self.perfumes = unique_perfumes
         return unique_perfumes
     
-    def save_to_json(self, filename: str = 'aroma_euro_catalog.json') -> None:
+    def save_to_json(self, filename: str = 'aroma_euro_perfumes.json') -> None:
         """Сохранение данных в JSON файл"""
         try:
             # Подготавливаем данные для сохранения
             save_data = {
                 'metadata': {
                     'source': 'aroma-euro.ru',
-                    'catalog_url': self.catalog_url,
                     'parsing_date': time.strftime('%Y-%m-%d %H:%M:%S'),
                     'total_count': len(self.perfumes),
-                    'parser_version': '3.0'
+                    'parser_version': 'final-2.0'
                 },
                 'perfumes': self.perfumes
             }
@@ -353,30 +369,21 @@ class AromaEuroFinalParser:
 
 def main():
     """Основная функция"""
-    parser = AromaEuroFinalParser()
-    
-    # Опция для тестирования на ограниченном количестве страниц
-    max_pages = None
-    if len(sys.argv) > 1:
-        try:
-            max_pages = int(sys.argv[1])
-            print(f"Парсинг ограничен до {max_pages} страниц")
-        except ValueError:
-            print("Неверный аргумент для количества страниц")
+    parser = AromaEuroParser()
     
     try:
         # Парсинг каталога
-        perfumes = parser.parse_all_catalog(max_pages=max_pages)
+        perfumes = parser.parse_all_catalog()
         
         if perfumes:
             # Сохранение в JSON
-            parser.save_to_json('aroma_euro_catalog.json')
+            parser.save_to_json('aroma_euro_perfumes.json')
             
             # Вывод статистики
             stats = parser.get_statistics()
-            print("\n" + "="*60)
+            print("\n" + "="*50)
             print("СТАТИСТИКА ПАРСИНГА AROMA-EURO.RU")
-            print("="*60)
+            print("="*50)
             print(f"Всего парфюмов: {stats.get('total_perfumes', 0)}")
             print(f"Уникальных брендов: {stats.get('unique_brands', 0)}")
             print(f"С определенным брендом: {stats.get('names_with_brands', 0)}")
@@ -386,9 +393,9 @@ def main():
             print(f"С изображениями: {stats.get('with_images', 0)}")
             
             # Показать первые несколько записей
-            print("\n" + "="*60)
+            print("\n" + "="*50)
             print("ПРИМЕРЫ ЗАПИСЕЙ")
-            print("="*60)
+            print("="*50)
             for i, perfume in enumerate(perfumes[:10], 1):
                 print(f"\n{i}. {perfume.get('full_title', 'Без названия')}")
                 if perfume.get('brand'):
@@ -398,7 +405,7 @@ def main():
                 if perfume.get('price'):
                     print(f"   Цена: {perfume['price']}")
                 if perfume.get('url'):
-                    print(f"   URL: {perfume['url'][:80]}...")
+                    print(f"   URL: {perfume['url']}")
             
             # Топ брендов
             brand_counts = {}
@@ -408,11 +415,11 @@ def main():
                     brand_counts[brand] = brand_counts.get(brand, 0) + 1
             
             if brand_counts:
-                print("\n" + "="*60)
-                print("ТОП-15 БРЕНДОВ")
-                print("="*60)
+                print("\n" + "="*50)
+                print("ТОП-10 БРЕНДОВ")
+                print("="*50)
                 sorted_brands = sorted(brand_counts.items(), key=lambda x: x[1], reverse=True)
-                for i, (brand, count) in enumerate(sorted_brands[:15], 1):
+                for i, (brand, count) in enumerate(sorted_brands[:10], 1):
                     print(f"{i:2d}. {brand}: {count} товаров")
         else:
             logger.warning("Не удалось извлечь данные о парфюмах")
